Separate out Processor from BlockProcessor
