+++ executor/src/witgen/block_processor.rs
@@ -1,46 +1,15 @@
-use std::{collections::HashSet, marker::PhantomData};
+use std::collections::HashSet;
 
-use ast::{
-    analyzed::{AlgebraicExpression as Expression, AlgebraicReference, Identity, PolyID},
-    parsed::SelectedExpressions,
-};
+use ast::analyzed::{AlgebraicExpression as Expression, AlgebraicReference, Identity, PolyID};
 use number::FieldElement;
 
-use crate::witgen::{query_processor::QueryProcessor, Constraint};
-
 use super::{
-    affine_expression::AffineExpression,
-    data_structures::{column_map::WitnessColumnMap, finalizable_data::FinalizableData},
-    identity_processor::IdentityProcessor,
-    rows::{RowFactory, RowPair, RowUpdater, UnknownStrategy},
+    data_structures::finalizable_data::FinalizableData,
+    processor::{Left, OuterQuery, Processor, WithCalldata, WithoutCalldata},
+    rows::RowFactory,
     sequence_iterator::{Action, ProcessingSequenceIterator, SequenceStep},
-    Constraints, EvalError, EvalValue, FixedData, MutableState, QueryCallback,
+    Constraints, EvalError, FixedData, MutableState, QueryCallback,
 };
-
-type Left<'a, T> = Vec<AffineExpression<&'a AlgebraicReference, T>>;
-
-// Marker types
-pub struct WithCalldata;
-pub struct WithoutCalldata;
-
-/// Data needed to handle an outer query.
-pub struct OuterQuery<'a, T: FieldElement> {
-    /// A local copy of the left-hand side of the outer query.
-    /// This will be mutated while processing the block.
-    pub left: Left<'a, T>,
-    /// The right-hand side of the outer query.
-    pub right: &'a SelectedExpressions<Expression<T>>,
-}
-
-impl<'a, T: FieldElement> OuterQuery<'a, T> {
-    pub fn new(left: Left<'a, T>, right: &'a SelectedExpressions<Expression<T>>) -> Self {
-        Self { left, right }
-    }
-
-    pub fn is_complete(&self) -> bool {
-        self.left.iter().all(|l| l.is_constant())
-    }
-}
 
 /// A basic processor that knows how to determine a unique satisfying witness
 /// for a given list of identities.
@@ -49,25 +18,7 @@ impl<'a, T: FieldElement> OuterQuery<'a, T> {
 /// - `'b`: The duration of this machine's call (e.g. the mutable references of the other machines)
 /// - `'c`: The duration of this Processor's lifetime (e.g. the reference to the identity processor)
 pub struct BlockProcessor<'a, 'b, 'c, T: FieldElement, Q: QueryCallback<T>, CalldataAvailable> {
-    /// The global index of the first row of [Processor::data].
-    row_offset: u64,
-    /// The rows that are being processed.
-    data: FinalizableData<'a, T>,
-    /// The list of identities
-    identities: &'c [&'a Identity<Expression<T>>],
-    /// The mutable state
-    mutable_state: &'c mut MutableState<'a, 'b, T, Q>,
-    /// The fixed data (containing information about all columns)
-    fixed_data: &'a FixedData<'a, T>,
-    /// The row factory
-    row_factory: RowFactory<'a, T>,
-    /// The set of witness columns that are actually part of this machine.
-    witness_cols: &'c HashSet<PolyID>,
-    /// Whether a given witness column is relevant for this machine (faster than doing a contains check on witness_cols)
-    is_relevant_witness: WitnessColumnMap<bool>,
-    /// The outer query, if any. If there is none, processing an outer query will fail.
-    outer_query: Option<OuterQuery<'a, T>>,
-    _marker: PhantomData<CalldataAvailable>,
+    processor: Processor<'a, 'b, 'c, T, Q, CalldataAvailable>,
 }
 
 impl<'a, 'b, 'c, T: FieldElement, Q: QueryCallback<T>>
@@ -82,13 +33,7 @@ impl<'a, 'b, 'c, T: FieldElement, Q: QueryCallback<T>>
         row_factory: RowFactory<'a, T>,
         witness_cols: &'c HashSet<PolyID>,
     ) -> Self {
-        let is_relevant_witness = WitnessColumnMap::from(
-            fixed_data
-                .witness_cols
-                .keys()
-                .map(|poly_id| witness_cols.contains(&poly_id)),
-        );
-        Self {
+        let processor = Processor::new(
             row_offset,
             data,
             mutable_state,
@@ -96,65 +41,33 @@ impl<'a, 'b, 'c, T: FieldElement, Q: QueryCallback<T>>
             fixed_data,
             row_factory,
             witness_cols,
-            is_relevant_witness,
-            outer_query: None,
-            _marker: PhantomData,
-        }
+        );
+        Self { processor }
     }
 
     pub fn with_outer_query(
         self,
         outer_query: OuterQuery<'a, T>,
     ) -> BlockProcessor<'a, 'b, 'c, T, Q, WithCalldata> {
-        BlockProcessor {
-            outer_query: Some(outer_query),
-            _marker: PhantomData,
-            row_offset: self.row_offset,
-            data: self.data,
-            mutable_state: self.mutable_state,
-            identities: self.identities,
-            fixed_data: self.fixed_data,
-            row_factory: self.row_factory,
-            witness_cols: self.witness_cols,
-            is_relevant_witness: self.is_relevant_witness,
-        }
+        let processor = self.processor.with_outer_query(outer_query);
+        BlockProcessor::<_, _, WithCalldata> { processor }
     }
 
     pub fn finish(self) -> FinalizableData<'a, T> {
-        self.data
+        self.processor.finish()
     }
 }
 
 impl<'a, 'b, T: FieldElement, Q: QueryCallback<T>> BlockProcessor<'a, 'b, '_, T, Q, WithCalldata> {
     /// Destroys itself, returns the data and updated left-hand side of the outer query (if available).
     pub fn finish(self) -> (FinalizableData<'a, T>, Left<'a, T>) {
-        (self.data, self.outer_query.unwrap().left)
+        self.processor.finish()
     }
 }
 
 impl<'a, 'b, T: FieldElement, Q: QueryCallback<T>, CalldataAvailable>
     BlockProcessor<'a, 'b, '_, T, Q, CalldataAvailable>
 {
-    /// Evaluate all identities on all *non-wrapping* row pairs, assuming zero for unknown values.
-    /// If any identity was unsatisfied, returns an error.
-    #[allow(dead_code)]
-    pub fn check_constraints(&mut self) -> Result<(), EvalError<T>> {
-        let mut identity_processor = IdentityProcessor::new(self.fixed_data, self.mutable_state);
-        for i in 0..(self.data.len() - 1) {
-            let row_pair = RowPair::new(
-                &self.data[i],
-                &self.data[i + 1],
-                self.row_offset + i as u64,
-                self.fixed_data,
-                UnknownStrategy::Zero,
-            );
-            for identity in self.identities {
-                identity_processor.process_identity(identity, &row_pair)?;
-            }
-        }
-        Ok(())
-    }
-
     /// Figures out unknown values.
     /// Returns the assignments to outer query columns.
     pub fn solve(
@@ -167,156 +80,19 @@ impl<'a, 'b, T: FieldElement, Q: QueryCallback<T>, CalldataAvailable>
             let row_index = (1 + row_delta) as usize;
             let progress = match action {
                 Action::InternalIdentity(identity_index) => {
-                    self.process_identity(row_index, identity_index)?
+                    self.processor.process_identity(row_index, identity_index)?
                 }
                 Action::OuterQuery => {
-                    let (progress, new_outer_assignments) = self.process_outer_query(row_index)?;
+                    let (progress, new_outer_assignments) =
+                        self.processor.process_outer_query(row_index)?;
                     outer_assignments.extend(new_outer_assignments);
                     progress
                 }
-                Action::ProverQueries => self.process_queries(row_index),
+                Action::ProverQueries => self.processor.process_queries(row_index),
             };
             sequence_iterator.report_progress(progress);
         }
         Ok(outer_assignments)
-    }
-
-    fn process_queries(&mut self, row_index: usize) -> bool {
-        let mut query_processor =
-            QueryProcessor::new(self.fixed_data, self.mutable_state.query_callback);
-        let global_row_index = self.row_offset + row_index as u64;
-        let row_pair = RowPair::new(
-            &self.data[row_index],
-            &self.data[row_index + 1],
-            global_row_index,
-            self.fixed_data,
-            UnknownStrategy::Unknown,
-        );
-        let mut updates = EvalValue::complete(vec![]);
-        for poly_id in self.fixed_data.witness_cols.keys() {
-            if self.is_relevant_witness[&poly_id] {
-                updates.combine(query_processor.process_query(&row_pair, &poly_id));
-            }
-        }
-        self.apply_updates(row_index, &updates, || "queries".to_string())
-    }
-
-    /// Given a row and identity index, computes any updates, applies them and returns
-    /// whether any progress was made.
-    fn process_identity(
-        &mut self,
-        row_index: usize,
-        identity_index: usize,
-    ) -> Result<bool, EvalError<T>> {
-        let identity = &self.identities[identity_index];
-
-        // Create row pair
-        let global_row_index = self.row_offset + row_index as u64;
-        let row_pair = RowPair::new(
-            &self.data[row_index],
-            &self.data[row_index + 1],
-            global_row_index,
-            self.fixed_data,
-            UnknownStrategy::Unknown,
-        );
-
-        // Compute updates
-        let mut identity_processor = IdentityProcessor::new(self.fixed_data, self.mutable_state);
-        let updates = identity_processor
-            .process_identity(identity, &row_pair)
-            .map_err(|e| {
-                log::warn!("Error in identity: {identity}");
-                log::warn!(
-                    "Known values in current row (local: {row_index}, global {global_row_index}):\n{}",
-                    self.data[row_index].render_values(false, Some(self.witness_cols)),
-                );
-                if identity.contains_next_ref() {
-                    log::warn!(
-                        "Known values in next row (local: {}, global {}):\n{}",
-                        row_index + 1,
-                        global_row_index + 1,
-                        self.data[row_index + 1].render_values(false, Some(self.witness_cols)),
-                    );
-                }
-                e
-            })?;
-
-        Ok(self.apply_updates(row_index, &updates, || identity.to_string()))
-    }
-
-    fn process_outer_query(
-        &mut self,
-        row_index: usize,
-    ) -> Result<(bool, Constraints<&'a AlgebraicReference, T>), EvalError<T>> {
-        let OuterQuery { left, right } = self
-            .outer_query
-            .as_mut()
-            .expect("Asked to process outer query, but it was not set!");
-
-        let row_pair = RowPair::new(
-            &self.data[row_index],
-            &self.data[row_index + 1],
-            self.row_offset + row_index as u64,
-            self.fixed_data,
-            UnknownStrategy::Unknown,
-        );
-
-        let mut identity_processor = IdentityProcessor::new(self.fixed_data, self.mutable_state);
-        let updates = identity_processor
-            .process_link(left, right, &row_pair)
-            .map_err(|e| {
-                log::warn!("Error in outer query: {e}");
-                log::warn!("Some of the following entries could not be matched:");
-                for (l, r) in left.iter().zip(right.expressions.iter()) {
-                    if let Ok(r) = row_pair.evaluate(r) {
-                        log::warn!("  => {} = {}", l, r);
-                    }
-                }
-                e
-            })?;
-
-        let progress = self.apply_updates(row_index, &updates, || "outer query".to_string());
-
-        let outer_assignments = updates
-            .constraints
-            .into_iter()
-            .filter(|(poly, _)| !self.witness_cols.contains(&poly.poly_id))
-            .collect::<Vec<_>>();
-
-        Ok((progress, outer_assignments))
-    }
-
-    fn apply_updates(
-        &mut self,
-        row_index: usize,
-        updates: &EvalValue<&'a AlgebraicReference, T>,
-        source_name: impl Fn() -> String,
-    ) -> bool {
-        if updates.constraints.is_empty() {
-            return false;
-        }
-
-        log::trace!("    Updates from: {}", source_name());
-
-        // Build RowUpdater
-        // (a bit complicated, because we need two mutable
-        // references to elements of the same vector)
-        let (current, next) = self.data.mutable_row_pair(row_index);
-        let mut row_updater = RowUpdater::new(current, next, self.row_offset + row_index as u64);
-
-        for (poly, c) in &updates.constraints {
-            if self.witness_cols.contains(&poly.poly_id) {
-                row_updater.apply_update(poly, c);
-            } else if let Constraint::Assignment(v) = c {
-                let left = &mut self.outer_query.as_mut().unwrap().left;
-                log::trace!("      => {} (outer) = {}", poly, v);
-                for l in left.iter_mut() {
-                    l.assign(poly, *v);
-                }
-            };
-        }
-
-        true
     }
 }
 
@@ -359,7 +135,7 @@ mod tests {
     fn do_with_processor<T: FieldElement, Q: QueryCallback<T>, R>(
         src: &str,
         mut query_callback: Q,
-        f: impl Fn(&mut BlockProcessor<T, Q, WithoutCalldata>, BTreeMap<String, PolyID>) -> R,
+        f: impl Fn(BlockProcessor<T, Q, WithoutCalldata>, BTreeMap<String, PolyID>, u64, usize) -> R,
     ) -> R {
         let analyzed = analyze_string(src);
         let (constants, degree) = generate(&analyzed);
@@ -396,7 +172,7 @@ mod tests {
         let identities = analyzed.identities.iter().collect::<Vec<_>>();
         let witness_cols = fixed_data.witness_cols.keys().collect();
 
-        let mut processor = BlockProcessor::new(
+        let processor = BlockProcessor::new(
             row_offset,
             data,
             &mut mutable_state,
@@ -406,31 +182,36 @@ mod tests {
             &witness_cols,
         );
 
-        f(&mut processor, name_to_poly_id(&fixed_data))
+        f(
+            processor,
+            name_to_poly_id(&fixed_data),
+            degree,
+            identities.len(),
+        )
     }
 
     fn solve_and_assert<T: FieldElement>(src: &str, asserted_values: &[(usize, &str, u64)]) {
         let query_callback = |_: &str| -> Option<T> { None };
-        do_with_processor(src, query_callback, |processor, poly_ids| {
-            let mut sequence_iterator =
-                ProcessingSequenceIterator::Default(DefaultSequenceIterator::new(
-                    processor.data.len() - 2,
-                    processor.identities.len(),
-                    None,
-                ));
-            let outer_updates = processor.solve(&mut sequence_iterator).unwrap();
-            assert!(outer_updates.is_empty());
+        do_with_processor(
+            src,
+            query_callback,
+            |mut processor, poly_ids, degree, num_identities| {
+                let mut sequence_iterator = ProcessingSequenceIterator::Default(
+                    DefaultSequenceIterator::new(degree as usize - 2, num_identities, None),
+                );
+                let outer_updates = processor.solve(&mut sequence_iterator).unwrap();
+                assert!(outer_updates.is_empty());
 
-            // Can't use processor.finish(), because we don't own it...
-            let data = processor.data.clone();
+                let data = processor.finish();
 
-            for &(i, name, expected) in asserted_values.iter() {
-                let poly_id = poly_ids[name];
-                let row = &data[i];
-                let actual: T = row[&poly_id].value.unwrap_or_default();
-                assert_eq!(actual, T::from(expected));
-            }
-        })
+                for &(i, name, expected) in asserted_values.iter() {
+                    let poly_id = poly_ids[name];
+                    let row = &data[i];
+                    let actual: T = row[&poly_id].value.unwrap_or_default();
+                    assert_eq!(actual, T::from(expected));
+                }
+            },
+        )
     }
 
     #[test]
