+++ executor/src/witgen/mod.rs
@@ -28,6 +28,7 @@ mod generator;
 mod global_constraints;
 mod identity_processor;
 mod machines;
+mod processor;
 mod query_processor;
 mod range_constraints;
 mod rows;
